feat: [노트 상세보기] 서스펜스 적용
