+++ src/apis/note/fetchNoteDetail.ts
@@ -4,5 +4,8 @@ import instance from "../apiClient";
 
 export const fetchNoteDetail = async (
   noteId: number,
-): Promise<TeamIdNotesGet200ResponseNotesInner> =>
-  (await instance.get(`/notes/${noteId}`)).data;
+): Promise<TeamIdNotesGet200ResponseNotesInner | null> => {
+  if (noteId == null) return Promise.resolve(null);
+
+  return (await instance.get(`/notes/${noteId}`)).data;
+};
